Update for Krastorio2, and bugfix
I noticed this mod wasn't working when Krastorio2 + SE were both installed, because they also add space-science-pack and se-rocket-science pack.  

Also, there was a bug when there was more than one ingredient: when you call table.remove(..., i), it changes the indexes of all the following items, so the next iteration of 'i' is off by one, making it skip the next ingredient.  I fixed this by simply doing starting the for loops from the end of the list.